Replace old plasmod Solver with refactored
