--- bluemira/codes/plasmod/PLASMOD_DEFAULT_IN.json
@@ -1,117 +1,0 @@
-{
-    "tol": 1e-10,
-    "dtmin": 1e-2,
-    "dtmax": 1e-2,
-    "dtmaxmin": 0.1,
-    "dtmaxmax": 1.0,
-    "dtminmax": 5.0,
-    "dt": 0.01,
-    "dtinc": 2.0,
-    "Ainc": 1.1,
-    "test": 10000,
-    "tolmin": 10.1,
-    "dgy": 1e-5,
-    "eopt": 0.1,
-    "capA": 0.1,
-    "i_diagz": 0,
-    "isiccir": 0,
-    "isawt": 1,
-    "nx": 41,
-    "nxt": 5,
-    "nchannels": 3,
-    "i_impmodel": 1,
-    "i_modeltype": 1,
-    "i_equiltype": 2,
-    "i_pedestal": 2,
-    "ntglf": 11,
-    "xtglf_1": 0.1,
-    "xtglf_2": 0.15,
-    "xtglf_3": 0.2,
-    "xtglf_4": 0.25,
-    "xtglf_5": 0.3,
-    "xtglf_6": 0.4,
-    "xtglf_7": 0.5,
-    "xtglf_8": 0.6,
-    "xtglf_9": 0.7,
-    "xtglf_10": 0.75,
-    "xtglf_11": 0.8,
-    "R": 9.002,
-    "Ip": 17.75,
-    "A": 3.1,
-    "volume_in": -2500,
-    "amin": 2.9039,
-    "Bt": 5.855,
-    "d": 0.38491934960310104,
-    "d95": 0.333,
-    "k": 1.6969830041844367,
-    "k95": 1.652,
-    "q95": 3.5,
-    "fuelmix": 0.5,
-    "fuelhe3": 0.0,
-    "globtau_d": 10.0,
-    "globtau_t": 10.0,
-    "globtau_he": 10.0,
-    "globtau_xe": 10.0,
-    "globtau_ar": 10.0,
-    "car": 0.0,
-    "cxe": 0.0,
-    "che": 0.0,
-    "che3": 0.0,
-    "cprotium": 0.0,
-    "cwol": 0.0,
-    "psepplh_inf": 0.1,
-    "psepplh_sup": 1000.0,
-    "pradpos": 0.7,
-    "pradfrac": 0.6,
-    "psepb_q95AR_sup": 9.2,
-    "psep_r_sup": 230.0,
-    "fcoreraditv": -1.0,
-    "qdivt_sup": 0.0,
-    "c_car": 10.0,
-    "rho_n": 0.94,
-    "rho_T": 0.94,
-    "teped": 5.5,
-    "tesep": 0.1,
-    "pedscal": 1.0,
-    "f_gw": 0.85,
-    "f_gws": 0.5,
-    "Hfact": 1.1,
-    "fpion": 0.5,
-    "nbcdeff": 0.3,
-    "eccdeff": 0.3,
-    "x_control_nbi": 0.0,
-    "x_control_ech": 0.0,
-    "dx_control_nbi": 0.2,
-    "dx_control_ech": 0.03,
-    "x_cd_nbi": 0.0,
-    "x_cd_ech": 0.0,
-    "dx_cd_nbi": 0.2,
-    "dx_cd_ech": 0.03,
-    "x_fus_nbi": 0.0,
-    "x_fus_ech": 0.0,
-    "dx_fus_nbi": 0.2,
-    "dx_fus_ech": 0.03,
-    "x_heat_nbi": 0.0,
-    "x_heat_ech": 0.0,
-    "dx_heat_nbi": 0.2,
-    "dx_heat_ech": 0.03,
-    "nbi_energy": 1000.0,
-    "pfus_req": 0.0,
-    "f_ni": 0.0,
-    "pheat_max": 130.0,
-    "q_control": 130.0,
-    "q_heat": 0.0,
-    "q_cd": 0.0,
-    "q_fus": 0.0,
-    "pech": 0.0,
-    "pnbi": 0.0,
-    "fcdp": -1.0,
-    "maxpauxor": 20.0,
-    "v_loop": -1.0e-6,
-    "plh": 6,
-    "qnbi_psepfac": 100.0,
-    "cxe_psepfac": 1.0e-3,
-    "car_qdivt": 1.0e-4,
-    "contrpovs": 0.0,
-    "contrpovr": 0.0
-}
