--- bluemira/codes/plasmod/PLASMOD_DEFAULT_OUT.json
@@ -1,98 +1,0 @@
-{
-    "amin": null,
-    "d": null,
-    "k": null,
-    "perim": null,
-    "bpolavg": null,
-    "betator": null,
-    "betapol": null,
-    "betan": null,
-    "f_gwpedtop": null,
-    "fbs": null,
-    "fcd": null,
-    "f_ni": null,
-    "Ip": null,
-    "q95": null,
-    "q_sep": null,
-    "qstar": null,
-    "rli": null,
-    "Sp": null,
-    "torsurf": null,
-    "Vp": null,
-    "v_loop": null,
-    "Hfact": null,
-    "Hcorr": null,
-    "taueff": null,
-    "tauee": null,
-    "tauei": null,
-    "Wth": null,
-    "rplas": null,
-    "Pfusdt": null,
-    "Pfusdd": null,
-    "Pfus": null,
-    "Pneut": null,
-    "Paux": null,
-    "Peaux": null,
-    "Piaux": null,
-    "Palpha": null,
-    "Prad": null,
-    "Pradcore": null,
-    "Pradedge": null,
-    "Psep": null,
-    "Psync": null,
-    "Pbrehms": null,
-    "Pline": null,
-    "PLH": null,
-    "Pohm": null,
-    "qcd": null,
-    "qheat": null,
-    "qfus": null,
-    "qdivt": null,
-    "psep_r": null,
-    "psepb_q95AR": null,
-    "cprotium": null,
-    "cwol": null,
-    "che": null,
-    "che3": null,
-    "cxe": null,
-    "car": null,
-    "Zeff": null,
-    "nped": null,
-    "nsep": null,
-    "teped": null,
-    "i_flag": null,
-    "jiter": null,
-    "av_ni": null,
-    "av_nd": null,
-    "av_nz": null,
-    "av_nhe": null,
-    "av_Ti": null,
-    "av_Te": null,
-    "av_Ten": null,
-    "dprof": null,
-    "ffprime": null,
-    "g2": null,
-    "g3": null,
-    "ipol": null,
-    "jbs": null,
-    "jcd": null,
-    "jpar": null,
-    "kprof": null,
-    "pprime": null,
-    "ndeut": null,
-    "ne": null,
-    "nfuel": null,
-    "nalf": null,
-    "nions": null,
-    "ntrit": null,
-    "phi": null,
-    "press": null,
-    "psi": null,
-    "qprof": null,
-    "shif": null,
-    "Te": null,
-    "Ti": null,
-    "x": null,
-    "volprof": null,
-    "vprime": null
-}
